feat(wip)!: UI overhaul and Next.js 15 migration
Switched to Next.js 15 and ShadCN UI components for a more modern and
polished design. Currently, the dashboard is a work in progress with
only the package list page being partially functional (it's missing
filters for now)

Additionally, the authentication system has been revamped to
automatically login to all the servers in the list and provide switcher
to change the active server.

Signed-off-by: SoulHarsh007 <harsh.peshwani@outlook.com>