feat!: add log parser and API route and improve charts
Adds back log parser (CachyTerm) and an API route to fetch logs for a
specific package and march (architecture) combination.

Improves the chart components to display statistics cleanly and
reduce unnecessary re-renders.

Add parsing logic to `Repo Actions` action to allow rendering multiple
packages in an expanded view for the tanstack table.

Update header breadcrumbs to include the march and pkgbase on the logs
page.

Fix zod schema to use non-negative integers instead of positive integers
as the values can be zero.

Signed-off-by: SoulHarsh007 <harsh.peshwani@outlook.com>