feat: add audit logs page
Integrates audit logs API into the dashboard, allowing users to view and
filter audit logs by username. It also parses the audit log based on the
description and displays it in a user-friendly format.

Show loading screen when server is changed using the server switcher.

Fix statistics page not updating statistics when the active server
changes.

Signed-off-by: SoulHarsh007 <harsh.peshwani@outlook.com>